refactored formApp code into own file
